requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
tqdm
PyYAML
torch>=2.2.0
deepxde>=1.11.0
